updated example files

+++ examples/BUCK_VM_PPO/1_train/clean.bat
@@ -3,15 +3,14 @@ choice /c YN /m "Are you sure you want to delete all the simulation artifacts? (
 if errorlevel 2 goto end
 
 echo Deleting *.pth, *.sp, *.txt, and the 'gym' folder...
-del /Q *.pth
-del /Q *.sp
-del /Q *.txt
-
-:: Remove read-only, hidden, and system attributes in the 'gym' folder
-attrib -r -h -s gym\*.* /S /D
+if exist "*.pth" del /Q "*.pth"
+if exist "*.sp" del /Q "*.sp"
+if exist "*.txt" del /Q "*.txt"
+if exist "*.net" del /Q "*.net"
 
 :: Delete the 'gym' folder
-rmdir /S /Q gym
+if exist "gym" attrib -r -h -s gym\*.* /S /D
+if exist "gym" rmdir /S /Q gym
 
 echo Deletion complete.
 :end
